% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencyConv.R
\name{currencyConv}
\alias{currencyConv}
\title{Convert numeric column from one currency to another currency.}
\usage{
currencyConv(
  data,
  vars,
  source_currency_column = NULL,
  target_currency_column = NULL,
  source_currency = NULL,
  target_currency = "GBP"
)
}
\arguments{
\item{data}{A dataframe with values to be converted.}

\item{vars}{A list containing column names of source values.}

\item{source_currency_column}{A character vector of the source currency's column name. (No need to define if the \code{source_currency} is filled)}

\item{target_currency_column}{A character vector of the target currency's column name. (No need to define if the \code{target_currency} is filled)}

\item{source_currency}{A character vector defining the source currency.}

\item{target_currency}{A character vector defining the target currency. If empty, it will convert values to GBP.}
}
\value{
A dataframe.
}
\description{
Convert numeric column from one currency to another currency.
}
\examples{
x <- data.frame(
  var1 = c(100,200,300),
  var2 = c(200, 99, 350),
  source = c("USD","DKK","USD"),
  target = c("GBP","GBP","GBP")
)

currencyConv(x, "var1")
currencyConv(x, "var1", "source")
currencyConv(x, "var1",target_currency = "GBP")
}
